changed USER to CART. Sending TOKEN in AXIOS CALLS to retrieve endpoints that need user id. Implemented middleware in api/cart.js. Changed store/cart.js intial state from [] to {}.
