+++ client/Routes.js
@@ -9,21 +9,21 @@ import Cart from "./components/Cart";
 import Payment from "./components/Payment"
 import Confirm from "./components/OrderConfirmation"
 import { me } from "./store";
-import { fetchCart } from "./store/singleUser";
+import { fetchCart } from "./store/cart";
 
 // COMPONENT
  
 class Routes extends Component {
   componentDidMount() {
     this.props.loadInitialData();
-    console.log('auth id:', this.props.auth.id)
+    
     // this.props.loadCart(this.props.auth.id)
   }
 
   render() {
 
-    console.log("ROUTES PROPS RENDER:",this.props)
-    console.log("ROUTES STATE RENDER:",this.state)
+    // console.log("ROUTES PROPS RENDER:",this.props)
+    // console.log("ROUTES STATE RENDER:",this.state)
     const { isLoggedIn } = this.props;
 
     return (
@@ -31,9 +31,9 @@ class Routes extends Component {
         {isLoggedIn ? (
           <Switch>
             <Route path="/" exact component={Home} />
-            <Route exact path="/flowers" render={(props)=> <AllFlowers {...props} />}/>
-            <Route path="/flowers/:id" render={()=><SingleFlower auth= {this.props.auth} user={()=> this.props.loadCart(this.props.auth.id)}/>} />
-            <Route path="/users/:userId/cart" render={()=><Cart auth= {this.props.auth} user={()=> this.props.loadCart(this.props.auth.id)}/>} />
+            <Route exact path="/flowers" component={AllFlowers}/>
+            <Route path="/flowers/:id" component={SingleFlower}/>
+            <Route path="/users/:userId/cart" component={cart}/>
             <Route path="/payment" component={Payment} />
             <Route path = "/confirmation" component={Confirm} />
 
@@ -60,7 +60,7 @@ const mapState = state => {
     // Otherwise, state.auth will be an empty object, and state.auth.id will be falsey
     isLoggedIn: !!state.auth.id,
     auth: state.auth,
-    user: state.user
+    cart: state.cart
   };
 };
