+++ client/store/user.js
@@ -1,0 +1,11 @@
+import axios from "axios";
+
+const GET_USER = "GET_USER";
+
+export const getUser = user => {
+    return {
+        type: GET_USER,
+        user
+    }
+}
+
