+++ server/api/gatekeepingMiddleware.js
@@ -1,5 +1,6 @@
 const { models: {User} } = require('../db')
 
+//intercepts token from axios call and spits out the user obj
 const requireToken = async ( req, res,next ) => {
     try {
         const token = req.headers.authorization
