+++ server/api/index.js
@@ -1,7 +1,7 @@
 const router = require('express').Router()
 module.exports = router
 
-router.use('/users', require('./users'))
+router.use('/cart', require('./cart'))
 router.use('/flowers', require('./flowers'))
